+++ R/plotMutationDensities.R
@@ -21,7 +21,7 @@
 #' @param mut.show.realtime logical; if `TRUE`, displays weeks post-conception
 #' on the evolutionary timeline.
 #' @param mut.snv.rate optional; rate of accumulated SNVs per day in a
-#' diploid genome (i.e. 3.2 SNVs/day in neuroblastoma)
+#' diploid genome (e.g., 3.2 SNVs/day in neuroblastoma).
 #' @param output.file optional; will save the plot.
 #' @param ref.build Reference genome. Default `hg19`.
 #' Can be `hg18`, `hg19` or `hg38`.
@@ -47,7 +47,7 @@
 #' plotMutationDensities(mrca)
 #' @export
 #' @importFrom graphics abline Axis box grid hist mtext par rect text title
-#' arrows legend points polygon segments
+#' @importFrom graphics arrows legend points polygon segments
 #' @importFrom stats setNames
 
 plotMutationDensities <- function(mrcaObj = NULL, samp.name = NULL,
@@ -56,7 +56,8 @@ plotMutationDensities <- function(mrcaObj = NULL, samp.name = NULL,
                                   mut.col.multi = "#176A02", mut.border = NULL,
                                   mut.show.density = TRUE, mut.breaks = NULL,
                                   mut.xaxis = NULL, mut.show.realtime = FALSE,
-                                  mut.snv.rate = 3.2, output.file = NULL, ...) {
+                                  mut.snv.rate = 3.2,
+                                  output.file = NULL, ...) {
     Seglength <- . <- A <- B <- variable <- value <- lines <- density <-
         chrom <- TCN <- Seglength <- n_mut_A <- n_mut_B <- n_mut_total_clonal <-
         density_total_mean <- density_A_mean <- density_B_mean <-
@@ -213,11 +214,22 @@ plotMutationDensities <- function(mrcaObj = NULL, samp.name = NULL,
     mtext(text = "SNVs per Mb", side = 1, line = 2, cex = 0.7)
 
     if (mut.show.realtime) {
-        weeks_pc <- c(12, 27, 38, 64, 90, 116)
+        # the maximal time in units of six months:
+        x.max.six.months <- x.max / mut.snv.rate * 2 * 3300 / (6 * 30)
+
+        # step size such that there are maximally 10 ticks
+        step.size <- ifelse(x.max.six.months > 10, 52, 26)
+        weeks_pc <- c(0, 12, 27, 38, 64, seq(
+            2 * 26, x.max.six.months * 26,
+            step.size
+        ) + 38)
         # Converting SNVs per day to SNVs per Mb starting from
         # gastrulation (-2 weeks), assuming haploid genome of 3300Mb
         snvs_per_mb <- (weeks_pc - 2) * 7 * mut.snv.rate / (3300 * 2)
-        realtime_labels <- c("12wk", "27wk", "38wk", "6mo", "12mo", "18mo")
+        realtime_labels <- c(
+            "", "12wk", "27wk", "38wk", "6mo",
+            paste(seq(2, x.max.six.months, 1 / (26 / step.size)) * 6, "mo", sep = "")
+        )
         axis(
             side = 3, at = c(x.min, snvs_per_mb, x.max),
             labels = c("", realtime_labels, ""),